Fix #62: restore right-click context menu on sidebar transcript rows
The sidebar's inline recording sub-rows (`RecordingSubRow`, shown when
"All Transcriptions" is expanded) had no `.contextMenu`, so right-click
did nothing — a regression versus the All Transcripts list, which still
worked. Users couldn't Rename / Move to Folder / Re-transcribe / Delete a
recording directly from the sidebar.

The two lists each built their own menu, which is how they drifted. Extract
the per-recording context menu — plus its sheets (rename, new folder,
send-to-LLM), permanent-delete confirmation, and helpers — into a single
`.recordingContextMenu(recording:selection:)` view modifier
(`RecordingContextMenu.swift`). `HistoryRow` and `RecordingSubRow` both apply
it, so the menus can't drift again.

Add a regression UI test (`OrganizationUITests`) that right-clicks the seeded
sidebar sub-row and asserts the context menu appears.
